remove os id scans from cc pipeline template
