+++ concourse/model/traits/__init__.py
@@ -15,7 +15,6 @@ def _traits():
     from .component_descriptor import ComponentDescriptorTrait
     from .cronjob import CronTrait
     from .draft_release import DraftReleaseTrait
-    from .image_scan import ImageScanTrait
     from .notifications import NotificationsTrait
     from .options import OptionsTrait
     from .publish import PublishTrait
@@ -29,7 +28,6 @@ def _traits():
         'component_descriptor': ComponentDescriptorTrait,
         'cronjob': CronTrait,
         'draft_release': DraftReleaseTrait,
-        'image_scan': ImageScanTrait,
         'notifications': NotificationsTrait,
         'options': OptionsTrait,
         'publish': PublishTrait,
