+++ delivery/model.py
@@ -2,7 +2,6 @@ import dataclasses
 import datetime
 import enum
 
-import awesomeversion
 import dacite
 import dateutil.parser
 
@@ -11,14 +10,6 @@ def _parse_datetime_if_present(date: str):
     if not date:
         return None
     return dateutil.parser.isoparse(date)
-
-
-def _parse_date_if_present(date: str | bool) -> datetime.date | bool | None:
-    if isinstance(date, bool):
-        return date
-    if not date:
-        return None
-    return dateutil.parser.isoparse(date).date()
 
 
 @dataclasses.dataclass # TODO: deduplicate w/ modelclass in delivery-service
@@ -61,28 +52,6 @@ class Sprint:
             config=dacite.Config(
                 type_hooks={datetime.datetime: _parse_datetime_if_present},
                 cast=[frozenset],
-            ),
-        )
-
-
-@dataclasses.dataclass(frozen=True)
-class OsReleaseInfo:
-    name: str
-    reached_eol: bool
-    greatest_version: str | None = None
-    eol_date: datetime.date | bool | None = None
-
-    @property
-    def parsed_version(self) -> awesomeversion.AwesomeVersion:
-        return awesomeversion.AwesomeVersion(self.name)
-
-    @staticmethod
-    def from_dict(raw: dict):
-        return dacite.from_dict(
-            data_class=OsReleaseInfo,
-            data=raw,
-            config=dacite.Config(
-                type_hooks={datetime.date | bool | None: _parse_date_if_present},
             ),
         )
